// Projects/Tamin/Eservices-UI/src/app/models/insurance-list/insurance.model.ts
export class InsuranceModel {
    approveDate: string;
    branchCode: string;
    brchCode: string;
    cityOfBirth: string;
    cityOfIssue: string;
    createdBy: string;
    creationTime: string;
    customerType: string;
    dateOfBirth: string;
    expCityCode: {
        cityCode:string
        cityCapital:string
        cityCodeOld:string
        cityDstnc:string
        cityGrade:string
        cityGroup:string
        cityName:string
        provinceCode:string
    };
    fatherName: string;
    finalTaeed: string;
    firstName: string;
    fisuId: string;
    flagNoTaeed2: string;
    flagPrint: string;
    flagSabt2: string;
    flagTaeed: string;
    flagTaeed2: string;
    id: string;
    idCardNumber: string;
    idCardSerial1: string;
    idCardSerial2: string;
    isuStatCode: string;
    isuTypeCode: string;
    lastName: string;
    natcin: string;
    nationalId: string;
    recogMethodDate: string;
    registerDate: string;
    requetNumber: string;
    ssn: string;
    gender: {
        genderCode: string
        genderDesc: string
        status: string
        statusDate: string
    };
    nation: {
        nationCode: string
        nationDesc: string
        status: string
        statusDate: string
    };
}